+++ test/main.hs
@@ -1,5 +1,5 @@
 -- Ignore warnings about using deprecated byLabel/fileByLabel functions
-{-# OPTIONS_GHC -fno-warn-warnings-deprecations #-}
+{-# OPTIONS_GHC -Wno-deprecations #-}
 
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -47,6 +47,8 @@ import Data.Maybe (isNothing)
 import qualified Data.Text as T
 import Yesod.Test.Internal (contentTypeHeaderIsUtf8)
 
+import qualified NestedRouteDispatchSpec.Foo.HandlerSpec
+
 parseQuery_ :: Text -> [[SelectorGroup]]
 parseQuery_ = either error id . parseQuery
 
@@ -67,6 +69,7 @@ mkYesod "RoutedApp" [parseRoutes|
 
 main :: IO ()
 main = hspec $ do
+    describe "NestedRouteDispatchSpec.Foo.HandlerSpec" NestedRouteDispatchSpec.Foo.HandlerSpec.spec
     describe "CSS selector parsing" $ do
         it "elements" $ parseQuery_ "strong" @?= [[DeepChildren [ByTagName "strong"]]]
         it "child elements" $ parseQuery_ "strong > i" @?= [[DeepChildren [ByTagName "strong"], DirectChildren [ByTagName "i"]]]
@@ -456,9 +459,11 @@ main = hspec $ do
             statusIs 201
 
             loc <- getLocation
-            liftIO $ assertBool "expected location to be available" $ isRight loc
-            let (Right (ResourceR t)) = loc
-            liftIO $ assertBool "expected location header to contain post param" $ t == "bar"
+            case loc of
+                Right (ResourceR t) ->
+                    liftIO $ assertBool "expected location header to contain post param" $ t == "bar"
+                other ->
+                    liftIO $ assertFailure $ "expected a Right ResourceR location, got: " <> show other
 
         yit "returns a Left when no redirect was returned" $ do
             get HomeR
@@ -497,11 +502,11 @@ main = hspec $ do
         yit "checks for valid content-type" $ do
             get ("get-json-wrong-content-type" :: Text)
             statusIs 200
-            (requireJSONResponse :: YesodExample site [Integer]) `liftedShouldThrow` (\(e :: SomeException) -> True)
+            (requireJSONResponse :: YesodExample site [Integer]) `liftedShouldThrow` (\(_ :: SomeException) -> True)
         yit "checks for valid JSON parse" $ do
             get ("get-json-response" :: Text)
             statusIs 200
-            (requireJSONResponse :: YesodExample site [Text]) `liftedShouldThrow` (\(e :: SomeException) -> True)
+            (requireJSONResponse :: YesodExample site [Text]) `liftedShouldThrow` (\(_ :: SomeException) -> True)
 
 instance RenderMessage LiteApp FormMessage where
     renderMessage _ _ = defaultFormMessage
@@ -649,8 +654,8 @@ getResourceR i = defaultLayout
 
 getIntegerR :: Handler Text
 getIntegerR = do
-    app <- getYesod
-    pure $ T.pack $ show (routedAppInteger app)
+    app' <- getYesod
+    pure $ T.pack $ show (routedAppInteger app')
 
 
 -- infix Copied from HSpec's version